Refactor: 거리계산 함수 모듈화
